+++ src/main/kotlin/godot/entrygenerator/filebuilder/EntryFileBuilderProvider.kt
@@ -5,13 +5,6 @@ import org.jetbrains.kotlin.resolve.BindingContext
 
 object EntryFileBuilderProvider {
 
-    fun provideEntryFileBuilder(generationType: EntryGenerationType, bindingContext: BindingContext): EntryFileBuilder {
-        return when(generationType) {
-            EntryGenerationType.KOTLIN_NATIVE -> KotlinNativeEntryFileBuilder(bindingContext)
-            EntryGenerationType.JVM -> JvmEntryFileBuilder(bindingContext)
-        }
-    }
-
     fun provideMainEntryFileBuilder(generationType: EntryGenerationType, bindingContext: BindingContext): EntryFileBuilder {
         return when(generationType) {
             EntryGenerationType.KOTLIN_NATIVE -> KotlinNativeEntryFileBuilder(bindingContext)
